move collision detection to its own file
